+++ app/models/extensions/ce_registry_resources.rb
@@ -1,7 +1,7 @@
 require 'active_support/concern'
 
-# CredentialRegistry specific behavior for resource envelopes
-module CredentialRegistryResources
+# CE/Registry specific behavior for resource envelopes
+module CERegistryResources
   extend ActiveSupport::Concern
 
   included do
@@ -9,10 +9,10 @@ module CredentialRegistryResources
       where('processed_resource @> ?', { 'ctdl:ctid' => ctid }.to_json)
     end)
 
-    validate :unique_ctid, if: :credential_registry?
+    validate :unique_ctid, if: :ce_registry?
 
     def unique_ctid
-      if Envelope.in_community('credential_registry')
+      if Envelope.in_community('ce_registry')
                  .where.not(envelope_id: envelope_id)
                  .with_ctid(processed_resource['ctdl:ctid'])
                  .exists?
@@ -20,8 +20,8 @@ module CredentialRegistryResources
       end
     end
 
-    def credential_registry?
-      community_name == 'credential_registry'
+    def ce_registry?
+      community_name =~ /ce_registry/
     end
 
     def self.generate_ctid