+++ db/seeds.rb
@@ -1,5 +1,5 @@
 EnvelopeCommunity.create([{ name: 'learning_registry',
                             backup_item: 'learning-registry-test',
                             default: true },
-                          { name: 'credential_registry',
-                            backup_item: 'credential-registry-test' }])
+                          { name: 'ce_registry',
+                            backup_item: 'ce-registry-test' }])